move repo() to conftest.py
