+++ tests/db/gitb/conftest.py
@@ -1,7 +1,7 @@
 from pathlib import Path
+import git
 
 import pytest
-
 
 ##__________________________________________________________________||
 @pytest.fixture()
@@ -21,5 +21,20 @@ def folder(tmpdir_factory):
     file1.write_text('abc\n')
     file2.write_text('123\n')
     yield folder
+
+##__________________________________________________________________||
+@pytest.fixture()
+def repo(folder):
+    """a git repo
+
+    The folder is initialized as a git repo.
+    The two files in the folder are committed.
+
+
+    """
+    repo = git.Repo.init(folder)
+    repo.git.add(A=True)
+    repo.index.commit('initial commit')
+    yield repo
 
 ##__________________________________________________________________||
